NEWSTART-554 Handle app URL config consistently
Happened when changing the app deployment from production to any other
deployment without restarting in between.  Not a big deal as this
functionality isn't in release builds, but fixed by ensuring we always
access the current deployment URL in the same way (that updates
whenever the preference is changed).

Also a few minor code style tweaks, and removed some unnecessary uses of
Optional.